Moves acme HTTP solver into shared /pkg/acme/httpsolver package
