+++ cmd/acmesolver/app/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/cmd/acmesolver/app",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/controller/acmechallenges/http/solver:go_default_library",
+        "//pkg/acme/httpsolver:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
