+++ pkg/acme/httpsolver/solver.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package solver
+package httpsolver
 
 import (
 	"fmt"
