+++ pkg/controller/acmechallenges/http/BUILD.bazel
@@ -11,10 +11,10 @@ go_library(
     importpath = "github.com/jetstack/cert-manager/pkg/controller/acmechallenges/http",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/acme/httpsolver:go_default_library",
         "//pkg/apis/acme/v1:go_default_library",
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/controller:go_default_library",
-        "//pkg/controller/acmechallenges/http/solver:go_default_library",
         "//pkg/logs:go_default_library",
         "//pkg/util:go_default_library",
         "@io_k8s_api//core/v1:go_default_library",
@@ -66,10 +66,7 @@ filegroup(
 
 filegroup(
     name = "all-srcs",
-    srcs = [
-        ":package-srcs",
-        "//pkg/controller/acmechallenges/http/solver:all-srcs",
-    ],
+    srcs = [":package-srcs"],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
 )
